fortune-teller/requirements.txt
Flask>=2.3,<4
pyserial==3.5
